Move profiling handlers to a new package perf
